+++ tosca/constants.go
@@ -25,3 +25,15 @@ const RunnableRunOperationName = RunnableInterfaceName + ".run"
 
 // RunnableCancelOperationName is the fully qualified name of the Cancel operation
 const RunnableCancelOperationName = RunnableInterfaceName + ".cancel"
+
+// StandardInterfaceName is the fully qualified name of the Standard interface
+const StandardInterfaceName = "tosca.interfaces.node.lifecycle.standard"
+
+// StandardInterfaceShortName is the short name of the Standard interface
+const StandardInterfaceShortName = "standard"
+
+// ConfigureInterfaceName is the fully qualified name of the Configure interface
+const ConfigureInterfaceName = "tosca.interfaces.relationships.configure"
+
+// ConfigureInterfaceShortName is the short name of the Configure interface
+const ConfigureInterfaceShortName = "configure"