+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+### BUG FIXES
+
+* Yorc should support long standard operation names as well as short ones ([GH-300](https://github.com/ystia/yorc/issues/300))
+
 ### ENHANCEMENTS
 
 * Bootstrap support of Ubuntu 1904 ([GH-419](https://github.com/ystia/yorc/issues/419))